# [pthread] Infinite Array Implementation