This PR replaces an inefficient linked list implementation with ESBMC’s infinite array model. 

- Each thread directly accesses its slot without iteration.
- Eliminates ESBMC’s verification overhead from dynamic memory allocation.

This PR gives one order of magnitude speedup for our regression tests that use `pthread_create_key'.

This PR (900s) uses the concurrency benchmarks: 

````
Statistics:           3179 Files
  correct:            1839
    correct true:     1404
    correct false:     435
  incorrect:            65
    incorrect true:     42
    incorrect false:    23
  unknown:            1275
  Score:              1531 (max: 5725)

https://github.com/esbmc/esbmc/actions/runs/13883021088
````

Master (900s) uses the concurrency benchmarks: 

````
Statistics:           3179 Files
  correct:            1839
    correct true:     1404
    correct false:     435
  incorrect:            65
    incorrect true:     42
    incorrect false:    23
  unknown:            1275
  Score:              1531 (max: 5725)

https://github.com/esbmc/esbmc/actions/runs/13884070327
````

This PR (30s):

````
Statistics:          33569 Files
  correct:           18070
    correct true:    10742
    correct false:    7328
  incorrect:            40
    incorrect true:     13
    incorrect false:    27
  unknown:           15459
  Score:             27964 (max: 55885)


https://github.com/esbmc/esbmc/actions/runs/13888616601
````

Master (30s):

````
Statistics:          33569 Files
  correct:           18077
    correct true:    10746
    correct false:    7331
  incorrect:            41
    incorrect true:     14
    incorrect false:    27
  unknown:           15451
  Score:             27943 (max: 55885)

GitHub Actions: https://github.com/esbmc/esbmc/actions/runs/13874366186
````