@lucasccordeiro 2025-03-18
Hi @intrigus-lgtm,

We have just run this PR for 30s to see whether we observe some regression.

We could increase the timeout (e.g., to 900 seconds) and see how many more benchmarks we could solve. However, with 900 seconds, the previous model could solve all verification tasks.

To observe the difference in performance, you could check the previous model and this proposed model using our regression suite. This is what I have done ;-)

Best,
Lucas