@intrigus-lgtm 2025-03-18
> Hi @intrigus-lgtm,
> 
> We have just run this PR for 30s to see whether we observe some regression.
> 
> We could increase the timeout (e.g., to 900 seconds) and see how many more benchmarks we could solve. However, with 900 seconds, the previous model could solve all verification tasks.
> 
> To observe the difference in performance, you could check the previous model and this proposed model using our regression suite. This is what I have done ;-)
> 
> Best, Lucas

I'm not doubting you, it's just that there is no real difference when comparing master and this pr :thinking: 
before-pr: https://github.com/esbmc/esbmc/actions/runs/13887098384/job/38853270771
pr: https://github.com/esbmc/esbmc/actions/runs/13897258794/job/38880605803

Anyway, it's good that this code has been removed as now TCs can be run with `--no-simplify` and don't need `--unwind` :)